+++ src/components/quiz-error/QuizError.scss
