+++ src/containers/App.jsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import { CSSTransition } from 'react-transition-group';
 
-import Loader from '../components/Loader/Loader';
-import Quiz from './Quiz/Quiz.jsx';
-import QuizError from '../components/QuizError/QuizError'
+import Loader from '../components/loader/Loader';
+import Quiz from './quiz/Quiz.jsx';
+import QuizError from '../components/quiz-error/QuizError'
 
 import { connect } from 'react-redux';
 import { bindActionCreators } from 'redux';
